consolidate pkg/content tests into one _test.go per source file
